// packages/client/src/generated/types.gen.ts
export type DuplicateBucketRequest = {
    icon?: string;
    name?: string;
};

export type EnqueuedReportResponse = {
    reportId?: number;
};

export type Exchange = {
    id?: number;
    name?: string;
    workingSchedules?: Array<WorkingSchedule>;
};

export type Fill = {
    filledAt?: string;
    id?: number;
    price?: number;
    quantity?: number;
    tradingMethod?: 'TOTV' | 'OTC';
    type?: 'TRADE' | 'STOCK_SPLIT' | 'STOCK_DISTRIBUTION' | 'FOP' | 'FOP_CORRECTION' | 'CUSTOM_STOCK_DISTRIBUTION' | 'EQUITY_RIGHTS' | 'SCRIP_STOCK_DIVIDENDS' | 'STOCK_DIVIDENDS' | 'STOCK_ACQUISITION' | 'CASH_AND_STOCK_ACQUISITION' | 'SPIN_OFF';
    walletImpact?: FillWalletImpact;
};

export type FillWalletImpact = {
    currency?: string;
    fxRate?: number;
    netValue?: number;
    realisedProfitLoss?: number;
    taxes?: Array<Tax>;
};

export type HistoricalOrder = {
    fill?: Fill;
    order?: Order;
};

export type HistoryDividendItem = {
    /**
     * In account's primary currency.
     */
    amount?: number;
    amountInEuro?: number;
    /**
     * The account's primary currency.
     */
    currency?: string;
    /**
     * In instrument currency
     */
    grossAmountPerShare?: number;
    instrument?: Instrument;
    paidOn?: string;
    quantity?: number;
    reference?: string;
    ticker?: string;
    tickerCurrency?: string;
    type?: 'ORDINARY' | 'BONUS' | 'PROPERTY_INCOME' | 'RETURN_OF_CAPITAL_NON_US' | 'DEMERGER' | 'INTEREST' | 'CAPITAL_GAINS_DISTRIBUTION_NON_US' | 'INTERIM_LIQUIDATION' | 'ORDINARY_MANUFACTURED_PAYMENT' | 'BONUS_MANUFACTURED_PAYMENT' | 'PROPERTY_INCOME_MANUFACTURED_PAYMENT' | 'RETURN_OF_CAPITAL_NON_US_MANUFACTURED_PAYMENT' | 'DEMERGER_MANUFACTURED_PAYMENT' | 'INTEREST_MANUFACTURED_PAYMENT' | 'CAPITAL_GAINS_DISTRIBUTION_NON_US_MANUFACTURED_PAYMENT' | 'INTERIM_LIQUIDATION_MANUFACTURED_PAYMENT' | 'INTEREST_PAID_BY_US_OBLIGORS' | 'INTEREST_PAID_BY_FOREIGN_CORPORATIONS' | 'DIVIDENDS_PAID_BY_US_CORPORATIONS' | 'DIVIDENDS_PAID_BY_FOREIGN_CORPORATIONS' | 'CAPITAL_GAINS' | 'REAL_PROPERTY_INCOME_AND_NATURAL_RESOURCES_ROYALTIES' | 'OTHER_INCOME' | 'QUALIFIED_INVESTMENT_ENTITY' | 'TRUST_DISTRIBUTION' | 'PUBLICLY_TRADED_PARTNERSHIP_DISTRIBUTION' | 'CAPITAL_GAINS_DISTRIBUTION' | 'RETURN_OF_CAPITAL' | 'OTHER_DIVIDEND_EQUIVALENT' | 'TAX_EVENT_1446F_FOR_PUBLICLY_TRADED_SECURITIES' | 'PTP_UNCHARACTERISED_INCOME' | 'MULTIPLE_1042S_TAX_COMPONENTS' | 'DIVIDEND' | 'SHORT_TERM_CAPITAL_GAINS' | 'LONG_TERM_CAPITAL_GAINS' | 'PROPERTY_INCOME_DISTRIBUTION' | 'TAX_EXEMPTED' | 'INTEREST_PAID_BY_US_OBLIGORS_MANUFACTURED_PAYMENT' | 'INTEREST_PAID_BY_FOREIGN_CORPORATIONS_MANUFACTURED_PAYMENT' | 'DIVIDENDS_PAID_BY_US_CORPORATIONS_MANUFACTURED_PAYMENT' | 'DIVIDENDS_PAID_BY_FOREIGN_CORPORATIONS_MANUFACTURED_PAYMENT' | 'CAPITAL_GAINS_MANUFACTURED_PAYMENT' | 'REAL_PROPERTY_INCOME_AND_NATURAL_RESOURCES_ROYALTIES_MANUFACTURED_PAYMENT' | 'OTHER_INCOME_MANUFACTURED_PAYMENT' | 'QUALIFIED_INVESTMENT_ENTITY_MANUFACTURED_PAYMENT' | 'TRUST_DISTRIBUTION_MANUFACTURED_PAYMENT' | 'PUBLICLY_TRADED_PARTNERSHIP_DISTRIBUTION_MANUFACTURED_PAYMENT' | 'CAPITAL_GAINS_DISTRIBUTION_MANUFACTURED_PAYMENT' | 'RETURN_OF_CAPITAL_MANUFACTURED_PAYMENT' | 'OTHER_DIVIDEND_EQUIVALENT_MANUFACTURED_PAYMENT' | 'TAX_EVENT_1446F_FOR_PUBLICLY_TRADED_SECURITIES_MANUFACTURED_PAYMENT' | 'PTP_UNCHARACTERISED_INCOME_MANUFACTURED_PAYMENT' | 'MULTIPLE_1042S_TAX_COMPONENTS_MANUFACTURED_PAYMENT' | 'DIVIDEND_MANUFACTURED_PAYMENT' | 'SHORT_TERM_CAPITAL_GAINS_MANUFACTURED_PAYMENT' | 'LONG_TERM_CAPITAL_GAINS_MANUFACTURED_PAYMENT' | 'PROPERTY_INCOME_DISTRIBUTION_MANUFACTURED_PAYMENT' | 'TAX_EXEMPTED_MANUFACTURED_PAYMENT';
};

export type HistoryTransactionItem = {
    /**
     * Amount in the currency of the transaction
     */
    amount?: number;
    /**
     * Currency of the transaction
     */
    currency?: string;
    dateTime?: string;
    /**
     * ID
     */
    reference?: string;
    type?: 'WITHDRAW' | 'DEPOSIT' | 'FEE' | 'TRANSFER' | 'INTEREST_ON_FREE_CASH' | 'LENDING_INTEREST';
};

/**
 * Instrument information as given by /instruments endpoint.
 */
export type Instrument = {
    /**
     * Instrument currency in ISO 4217 format.
     */
    currency?: string;
    /**
     * ISIN of the instrument.
     */
    isin?: string;
    /**
     * Name of the instrument.
     */
    name?: string;
    /**
     * Unique instrument identifier.
     */
    ticker?: string;
};
// ...
export type Order = {
    /**
     * The ISO 8601 formatted date of when the order was created.
     */
    createdAt?: string;
    /**
     * The currency used for the order in ISO 4217 format.
     */
    currency?: string;
    /**
     * If true, the order is eligible for execution outside regular trading hours.
     */
    extendedHours?: boolean;
    /**
     * The number of shares that have been successfully executed. Applicable to quantity orders.
     */
    filledQuantity?: number;
    /**
     * The total monetary value of the executed portion of the order. Applicable to orders placed by value.Note: Placing orders by value is not currently supported via the API but can be done through other Trading 212 platforms.
     */
    filledValue?: number;
    /**
     * A unique, system-generated identifier for the order.
     */
    id?: number;
    /**
     * How the order was initiated.
     */
    initiatedFrom?: 'API' | 'IOS' | 'ANDROID' | 'WEB' | 'SYSTEM' | 'AUTOINVEST' | 'INSTRUMENT_AUTOINVEST';
    instrument?: Instrument;
    /**
     * Applicable to LIMIT and STOP_LIMIT orders.
     */
    limitPrice?: number;
    /**
     * The total number of shares requested. Applicable to quantity orders.
     */
    quantity?: number;
    /**
     * Indicates whether the order is BUY or SELL.
     */
    side?: 'BUY' | 'SELL';
    /**
     * The current state of the order in its lifecycle.
     */
    status?: 'LOCAL' | 'UNCONFIRMED' | 'CONFIRMED' | 'NEW' | 'CANCELLING' | 'CANCELLED' | 'PARTIALLY_FILLED' | 'FILLED' | 'REJECTED' | 'REPLACING' | 'REPLACED';
    /**
     * Applicable to STOP and STOP_LIMIT orders.
     */
    stopPrice?: number;
    /**
     * The strategy used to place the order, either by QUANTITY or VALUE. The API currently only supports placing orders by QUANTITY.
     */
    strategy?: 'QUANTITY' | 'VALUE';
    /**
     * Unique instrument identifier. Get from the /instruments endpoint
     */
    ticker?: string;
    timeInForce?: TimeValidity;
    type?: 'LIMIT' | 'STOP' | 'MARKET' | 'STOP_LIMIT';
    /**
     * The total monetary value of the order. Applicable to value orders.
     */
    value?: number;
};

export type PaginatedResponseHistoricalOrder = {
    items?: Array<HistoricalOrder>;
    nextPagePath?: string | null;
};

export type PaginatedResponseHistoryDividendItem = {
    items?: Array<HistoryDividendItem>;
    nextPagePath?: string | null;
};

export type PaginatedResponseHistoryTransactionItem = {
    items?: Array<HistoryTransactionItem>;
    nextPagePath?: string | null;
};

export type PieRequest = {
    dividendCashAction?: 'REINVEST' | 'TO_ACCOUNT_CASH';
    endDate?: string;
    /**
     * Total desired value of the pie in account currency
     */
    goal?: number;
    icon?: string;
    instrumentShares?: {
        [key: string]: number;
    };
    name?: string;
};
// ...
export type TimeEvent = {
    date?: string;
    type?: 'OPEN' | 'CLOSE' | 'BREAK_START' | 'BREAK_END' | 'PRE_MARKET_OPEN' | 'AFTER_HOURS_OPEN' | 'AFTER_HOURS_CLOSE' | 'OVERNIGHT_OPEN';
};

/**
 * Specifies how long the order remains active:
 * * DAY: The order will automatically expire if not executed by midnight in the time zone of the instrument's exchange.
 * * GOOD_TILL_CANCEL: The order remains active indefinitely until it is either filled or explicitly cancelled by you.
 */
export type TimeValidity = 'DAY' | 'GOOD_TILL_CANCEL';

export type TradableInstrument = {
    /**
     * On the platform since
     */
    addedOn?: string;
    /**
     * ISO 4217
     */
    currencyCode?: string;
    extendedHours?: boolean;
    isin?: string;
    maxOpenQuantity?: number;
    name?: string;
    shortName?: string;
    /**
     * Unique identifier
     */
    ticker?: string;
    type?: 'CRYPTOCURRENCY' | 'ETF' | 'FOREX' | 'FUTURES' | 'INDEX' | 'STOCK' | 'WARRANT' | 'CRYPTO' | 'CVR' | 'CORPACT';
    /**
     * Get items in the /exchanges endpoint
     */
    workingScheduleId?: number;
};

export type WorkingSchedule = {
    id?: number;
    timeEvents?: Array<TimeEvent>;
};